suggestsrvchatgpt/requirements.txt
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
MarkupSafe==2.1.3
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
Werkzeug==2.3.6
openai==0.27.8
